Fixed some problems

+++ ProcessingGraph/types/PGImage_types.h
@@ -5,7 +5,7 @@
 typedef struct {
     int height, width;
     float * data; /* This can be directly cast to (PGColour_t *) */
-    float loc_x, loc_y; /* Where the image is, (0,0) by default */
+    int loc_x, loc_y; /* Where the image is, (0,0) by default */
 } PGImage_t;
 
 #endif
